requirements.txt
asgiref==3.7.2
bech32==1.2.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
dj-database-url==2.1.0
Django==5.0.1
django-bootstrap-static==4.2.1
dnspython==2.4.2
emoji==2.8.0
eva-decord==0.6.1
ffmpegio==0.8.5
ffmpegio-core==0.8.5
gunicorn==21.2.0
humanize==4.10.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
instaloader==4.10.1
lnurl==0.4.2
loguru==0.7.2
mongita==1.2.0
motor==3.5.1
moviepy==2.0.0.dev2
mypy-extensions==1.0.0
neo4j==5.20.0
nostr-sdk==0.33.0
numpy==1.26.3
packaging==23.2
pandas==2.1.3
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
proglog==0.1.10
psycopg2==2.9.9
pycryptodome==3.19.0
pydantic==1.10.13
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
PyUpload==0.1.4
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tabulate==0.9.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
zipp==3.17.0